+++ CHANGELOG.md
@@ -63,6 +63,7 @@
   * request broker annotation `@Request[Password]` now supports `minLength` attribute, default value is 8
   * request broker annotation `@Request[OneOf]` now throws an exception when list of allowed values is missing in annotation
   * changed `stubbles\input\web\WebRequest::protocolVersion()` to always report the protocol version when it can be detected, not just supported ones
+  * added `stubbles\input\web\WebRequest::originatingIpAddress()`
 
 
 2.5.1 (2014-05-07)